+++ moor/test/data/tables/todos.dart
@@ -71,6 +71,7 @@ class PureDefaults extends Table {
     'allTodosWithCategory': 'SELECT t.*, c.id as catId, c."desc" as catDesc '
         'FROM todos t INNER JOIN categories c ON c.id = t.category',
     'deleteTodoById': 'DELETE FROM todos WHERE id = ?',
+    'withIn': 'SELECT * FROM todos WHERE title = ?2 OR id IN ? OR title = ?1'
   },
 )
 class TodoDb extends _$TodoDb {