+++ moor_generator/lib/src/utils/string_escaper.dart
@@ -1,4 +1,11 @@
 String asDartLiteral(String value) {
-  final escaped = value.replaceAll("'", "\\'").replaceAll('\n', '\\n');
+  final escaped = escapeForDart(value);
   return "'$escaped'";
+}
+
+String escapeForDart(String value) {
+  return value
+      .replaceAll("'", "\\'")
+      .replaceAll('\$', '\\\$')
+      .replaceAll('\n', '\\n');
 }