+++ moor_generator/lib/src/writer/query_writer.dart
@@ -2,6 +2,7 @@ import 'package:moor_generator/src/model/specified_column.dart';
 import 'package:moor_generator/src/model/sql_query.dart';
 import 'package:moor_generator/src/utils/string_escaper.dart';
 import 'package:recase/recase.dart';
+import 'package:sqlparser/sqlparser.dart';
 
 /// Writes the handling code for a query. The code emitted will be a method that
 /// should be included in a generated database or dao class.
@@ -11,6 +12,18 @@ class QueryWriter {
   UpdatingQuery get _update => query as UpdatingQuery;
 
   QueryWriter(this.query);
+
+  /// The expanded sql that we insert into queries whenever an array variable
+  /// appears. For the query "SELECT * FROM t WHERE x IN ?", we generate
+  /// ```dart
+  /// test(List<int> var1) {
+  ///   final expandedvar1 = List.filled(var1.length, '?').join(',');
+  ///   customSelect('SELECT * FROM t WHERE x IN ($expandedvar1)', ...);
+  /// }
+  /// ```
+  String _expandedName(FoundVariable v) {
+    return 'expanded${v.dartParameterName}';
+  }
 
   void writeInto(StringBuffer buffer) {
     if (query is SqlSelectQuery) {
@@ -50,10 +63,11 @@ class QueryWriter {
   void _writeOneTimeReader(StringBuffer buffer) {
     buffer.write('Future<List<${_select.resultClassName}>> ${query.name}(');
     _writeParameters(buffer);
+    buffer.write(') {\n');
+    _writeExpandedDeclarations(buffer);
     buffer
-      ..write(') {\n')
       ..write('return (operateOn ?? this).') // use custom engine, if set
-      ..write('customSelect(${asDartLiteral(query.sql)},');
+      ..write('customSelect(${_queryCode()},');
     _writeVariables(buffer);
     buffer
       ..write(')')
@@ -70,9 +84,10 @@ class QueryWriter {
     // be used in transaction or similar context, only on the main database
     // engine.
     _writeParameters(buffer, dontOverrideEngine: true);
-    buffer
-      ..write(') {\n')
-      ..write('return customSelectStream(${asDartLiteral(query.sql)},');
+    buffer.write(') {\n');
+
+    _writeExpandedDeclarations(buffer);
+    buffer..write('return customSelectStream(${_queryCode()},');
 
     _writeVariables(buffer);
     buffer.write(',');
@@ -92,23 +107,29 @@ class QueryWriter {
      */
     buffer.write('Future<int> ${query.name}(');
     _writeParameters(buffer);
+    buffer.write(') {\n');
+
+    _writeExpandedDeclarations(buffer);
     buffer
-      ..write(') {\n')
       ..write('return (operateOn ?? this).')
-      ..write('customUpdate(${asDartLiteral(query.sql)},');
+      ..write('customUpdate(${_queryCode()},');
 
     _writeVariables(buffer);
     buffer.write(',');
     _writeUpdates(buffer);
 
-    buffer..write(');\n}\n');
+    buffer..write(',);\n}\n');
   }
 
   void _writeParameters(StringBuffer buffer,
       {bool dontOverrideEngine = false}) {
-    final paramList = query.variables
-        .map((v) => '${dartTypeNames[v.type]} ${v.dartParameterName}')
-        .join(', ');
+    final paramList = query.variables.map((v) {
+      var dartType = dartTypeNames[v.type];
+      if (v.isArray) {
+        dartType = 'List<$dartType>';
+      }
+      return '$dartType ${v.dartParameterName}';
+    }).join(', ');
 
     buffer.write(paramList);
 
@@ -120,16 +141,73 @@ class QueryWriter {
     }
   }
 
+  void _writeExpandedDeclarations(StringBuffer buffer) {
+    for (var variable in query.variables) {
+      if (variable.isArray) {
+        // final expandedvar1 = List.filled(var1.length, '?').join(',');
+        buffer
+          ..write('final ')
+          ..write(_expandedName(variable))
+          ..write(' = ')
+          ..write('List.filled(')
+          ..write(variable.dartParameterName)
+          ..write(".length, '?').join(',');");
+      }
+    }
+  }
+
   void _writeVariables(StringBuffer buffer) {
     buffer..write('variables: [');
 
     for (var variable in query.variables) {
-      buffer
-        ..write(createVariable[variable.type])
-        ..write('(${variable.dartParameterName}),');
+      // for a regular variable: Variable.withInt(x),
+      // for a list of vars: for (var $ in vars) Variable.withInt($),
+      final constructor = createVariable[variable.type];
+      final name = variable.dartParameterName;
+
+      if (variable.isArray) {
+        buffer.write('for (var \$ in $name) $constructor(\$)');
+      } else {
+        buffer.write('$constructor($name)');
+      }
+
+      buffer.write(',');
     }
 
     buffer..write(']');
+  }
+
+  /// Returns a Dart string literal representing the query after variables have
+  /// been expanded. For instance, 'SELECT * FROM t WHERE x IN ?' will be turned
+  /// into 'SELECT * FROM t WHERE x IN ($expandedVar1)'.
+  String _queryCode() {
+    // sort variables by the order in which they appear
+    final vars = query.fromContext.root.allDescendants
+        .whereType<Variable>()
+        .toList()
+          ..sort((a, b) => a.firstPosition.compareTo(b.firstPosition));
+
+    final buffer = StringBuffer("'");
+    var lastIndex = 0;
+
+    for (var sqlVar in vars) {
+      final moorVar = query.variables.singleWhere((f) => f.variable == sqlVar);
+      if (!moorVar.isArray) continue;
+
+      // write everything that comes before this var into the buffer
+      final currentIndex = sqlVar.firstPosition;
+      final queryPart = query.sql.substring(lastIndex, currentIndex);
+      buffer.write(escapeForDart(queryPart));
+      lastIndex = sqlVar.lastPosition;
+
+      // write the ($expandedVar) par
+      buffer.write('(\$${_expandedName(moorVar)})');
+    }
+
+    // write the final part after the last variable, plus the ending '
+    buffer..write(escapeForDart(query.sql.substring(lastIndex)))..write("'");
+
+    return buffer.toString();
   }
 
   void _writeReadsFrom(StringBuffer buffer) {